chore(planning): Initial work for plan optimizer
