+++ solver/src/prelude.rs
@@ -8,3 +8,6 @@ pub use crate::model::extensions::DomainsExt;
 pub use crate::model::lang::{IAtom, IVar};
 pub use crate::solver::SearchLimit;
 pub use crate::solver::Solver;
+
+pub use crate::core::INT_CST_MAX;
+pub use crate::core::INT_CST_MIN;
